+++ src/extension.ts
@@ -1,6 +1,8 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
+import { DOMParser, XMLSerializer } from '@xmldom/xmldom';
+import cronParser from 'cron-parser';
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -8,10 +10,58 @@ export function activate(context: vscode.ExtensionContext) {
 
 	const register = vscode.commands.registerTextEditorCommand;
 
+	class Result<T> {
+		private success: boolean;
+		private data: T;
+		private error: any;
+
+		constructor(_success: boolean, _data: T, _error: any) {
+			this.success = _success;
+			this.data = _data;
+			this.error = _error;
+		}
+
+		isOk(): boolean {
+			return this.success;
+		}
+
+		wrap(): T {
+			return this.data;
+		}
+
+		err(): any {
+			return this.error;
+		}
+
+		static succ<T>(_data: T): Result<T> {
+			return new Result(true, _data, undefined);
+		}
+
+		static fail<T>(_data: T): Result<T> {
+			return new Result(false, _data, undefined);
+		}
+
+		static err(_error: any): Result<any> {
+			return new Result(false, undefined, _error);
+		}
+
+	}
+
 	enum ConvertType {
 		UNKNOWN,
 		TIME,
 		UNICODE,
+		// *    *    *    *    *    *
+		// ┬    ┬    ┬    ┬    ┬    ┬
+		// │    │    │    │    │    |
+		// │    │    │    │    │    └ day of week (0 - 7, 1L - 7L) (0 or 7 is Sun)
+		// │    │    │    │    └───── month (1 - 12)
+		// │    │    │    └────────── day of month (1 - 31, L)
+		// │    │    └─────────────── hour (0 - 23)
+		// │    └──────────────────── minute (0 - 59)
+		// └───────────────────────── second (0 - 59, optional)
+		CRONTAB,
+		BASE64,
 	}
 
 	enum TimeUnitType {
@@ -36,7 +86,7 @@ export function activate(context: vscode.ExtensionContext) {
 
 	// Use the console to output diagnostic information (console.log) and errors (console.error)
 	// This line of code will only be executed once when your extension is activated
-	console.warn('"preview-conversion" active!');
+	// console.warn('"preview-conversion" active!');
 
 	const tokens = [
 		// vscode.commands.registerCommand('preview-conversion.conversion', () => {
@@ -49,9 +99,13 @@ export function activate(context: vscode.ExtensionContext) {
 			let fullText = document.getText();
 			fullText = convertUnicode(fullText);
 			fullText = fullText.replace(/-?\d{9}\d+/g, (_substring, _): string => {
-				let conversionResult = convertTime(_substring);
-				const dataOutput = new Date(Number(conversionResult.timestamp.toString()));
-				return formatDateByLocale(dataOutput);
+				const conversionResult = convertTime(_substring);
+				if (conversionResult.isOk()) {
+					const dataOutput = new Date(Number(conversionResult.wrap().timestamp.toString()));
+					return formatDateByLocale(dataOutput);
+				} else {
+					return _substring;
+				}
 			});
 			textEditor.edit((editBuilder) => {
 				editBuilder.replace(new vscode.Range(0, 0, document.lineCount, 0), fullText);
@@ -90,6 +144,9 @@ export function activate(context: vscode.ExtensionContext) {
 			// console.log('hoveredWord: ', range, '   ', hoveredWord);
 
 			const result = checkConvertTypeForMultiSource(hoveredWord);
+
+			// console.log('provideHover: ', result);
+
 			if (ConvertType.UNKNOWN === result.convertType) {
 				return undefined;
 			}
@@ -98,16 +155,30 @@ export function activate(context: vscode.ExtensionContext) {
 					return handleTime(hoveredWord[result.index]);
 				case ConvertType.UNICODE:
 					return handleUnicode(hoveredWord[result.index]);
+				case ConvertType.CRONTAB:
+					return handleCronTab(hoveredWord[result.index]);
+				case ConvertType.BASE64:
+					return handleBase64(hoveredWord[result.index]);
 			}
 		}
 	});
 
-	function convertTime(data: any): { cleanTime: any, timestamp: string, unitType: TimeUnitType, isHex: boolean, isFull: boolean, convertMillisecondsNotice: boolean, preGregorianCalendarNotice: boolean } {
-		console.log('convertTime:', data);
+	interface TimeData {
+		cleanTime: any,
+		timestamp: string,
+		unitType: TimeUnitType,
+		isHex: boolean,
+		isFull: boolean,
+		convertMillisecondsNotice: boolean,
+		preGregorianCalendarNotice: boolean,
+	}
 
-		const cleanTime = cleanTimestamp(data);
+	function convertTime(source: any): Result<TimeData> {
+		// console.log('convertTime:', data);
 
-		let result = {
+		const cleanTime = cleanTimestamp(source);
+
+		let data = {
 			cleanTime,
 			timestamp: '',
 			unitType: TimeUnitType.SECOND,
@@ -117,72 +188,72 @@ export function activate(context: vscode.ExtensionContext) {
 			preGregorianCalendarNotice: false,
 		};
 
-		if (cleanTime && cleanTime !== data.trim()) {
-			result.isFull = false;
+		if (cleanTime && cleanTime !== source.trim()) {
+			data.isFull = false;
 		}
 
 		if ((cleanTime.length === 0) || isNaN(cleanTime)) {
 			if (!isHex(cleanTime)) {
-				return result;
+				return Result.fail(data);
 			} else {
-				result.isHex = true;
+				data.isHex = true;
 			}
 		}
 
 		let _data = cleanTime * 1;
 		if ((_data >= 1E16) || (_data <= -1E16)) {
-			result.unitType = TimeUnitType.NANOSECONDS;
+			data.unitType = TimeUnitType.NANOSECONDS;
 			_data = Math.floor(_data / 1000000);
 		} else if ((_data >= 1E14) || (_data <= -1E14)) {
-			result.unitType = TimeUnitType.MICROSECONDS;
+			data.unitType = TimeUnitType.MICROSECONDS;
 			_data = Math.floor(_data / 1000);
 		} else if ((_data >= 1E11) || (_data <= -3E10)) {
-			result.unitType = TimeUnitType.MILLISECONDS;
+			data.unitType = TimeUnitType.MILLISECONDS;
 		} else {
-			result.unitType = TimeUnitType.SECOND;
+			data.unitType = TimeUnitType.SECOND;
 			if ((_data > 1E11) || (_data < -1E10)) {
-				result.convertMillisecondsNotice = true;
+				data.convertMillisecondsNotice = true;
 			}
 			_data = (_data * 1000);
 		}
 		if (_data < -68572224E5) {
-			result.preGregorianCalendarNotice = true;
+			data.preGregorianCalendarNotice = true;
 		}
 
-		result.timestamp = _data.toString();
+		data.timestamp = _data.toString();
 
-		return result;
+		return Result.succ(data);
 	}
 
 	function handleTime(hoveredWord: string): vscode.ProviderResult<vscode.Hover> {
 		let markdownArray: Array<vscode.MarkdownString> = [
-			new vscode.MarkdownString('# Conversion Timestamp', true)
+			new vscode.MarkdownString('**Conversion Timestamp**', true)
 		];
 
 		const conversionResult = convertTime(hoveredWord);
 
-		if (conversionResult.timestamp.length === 0) {
-			return;
-		} else {
-			if (!conversionResult.isFull) {
-				markdownArray.push(new vscode.MarkdownString(`Converting ${conversionResult.cleanTime} :`));
+		if (conversionResult.isOk()) {
+			let data = conversionResult.wrap();
+			if (!data.isFull) {
+				markdownArray.push(new vscode.MarkdownString(`Converting ${data.cleanTime} :`));
 			}
-			if (conversionResult.isHex) {
-				markdownArray.push(new vscode.MarkdownString(`Converting 0x${conversionResult.cleanTime} :`));
+			if (data.isHex) {
+				markdownArray.push(new vscode.MarkdownString(`Converting 0x${data.cleanTime} :`));
 			}
+			markdownArray.push(new vscode.MarkdownString(`检测到时间单位为${getTimeUnitTypeName(data.unitType)}:`));
+			markdownArray.push(new vscode.MarkdownString('----'));
 
-			markdownArray.push(new vscode.MarkdownString(`检测到时间单位为${getTimeUnitTypeName(conversionResult.unitType)}:`));
-			if (conversionResult.convertMillisecondsNotice) {
+			if (data.convertMillisecondsNotice) {
 				markdownArray.push(new vscode.MarkdownString('如果您尝试转换毫秒，请删除最后 3 位数字。'));
 			}
-			const dataOutput = new Date(Number(conversionResult.timestamp.toString()));
-			const gmt = new Date(Number(conversionResult.timestamp.toString())).toUTCString();
+			const dataOutput = new Date(Number(data.timestamp.toString()));
+			const gmt = new Date(Number(data.timestamp.toString())).toUTCString();
 			markdownArray.push(new vscode.MarkdownString('**GMT(标准时间)**'));
 			markdownArray.push(new vscode.MarkdownString(gmt));
 			markdownArray.push(new vscode.MarkdownString('**Your time zone(当前时区)**'));
 			markdownArray.push(new vscode.MarkdownString(dataOutput.toString()));
 			markdownArray.push(new vscode.MarkdownString(formatDateByLocale(dataOutput)));
-			if (conversionResult.preGregorianCalendarNotice) {
+			if (data.preGregorianCalendarNotice) {
 				markdownArray.push(new vscode.MarkdownString('1752 年 9 月 14 日（公历之前）之前的日期不准确。'));
 			}
 		}
@@ -190,6 +261,247 @@ export function activate(context: vscode.ExtensionContext) {
 		return {
 			contents: markdownArray
 		};
+	}
+
+	interface CronTabData {
+		timeList: string[]
+	}
+
+	function convertCronTab(hoveredWord: string): Result<CronTabData> {
+		/**
+		 * yyyy-MM-dd HH:mm:ss
+		 */
+		function formatDate(date: Date, withWeek: boolean = true): string {
+			let days = [
+				date.getFullYear(),
+				String(date.getMonth() + 1).padStart(2, '0'),
+				String(date.getDate()).padStart(2, '0'),
+			].join('-');
+			let times = [
+				String(date.getHours()).padStart(2, '0'),
+				String(date.getMinutes()).padStart(2, '0'),
+				String(date.getSeconds()).padStart(2, '0'),
+			].join(':');
+
+			let result = [days];
+			if (withWeek) {
+				const weekdays = ['星期天', '星期一', '星期二', '星期三', '星期四', '星期五', '星期六'];
+				result.push(weekdays[date.getDay()]);
+			}
+			result.push(times);
+
+			return result.join(' ');
+		}
+
+		let data: CronTabData = {
+			timeList: [],
+		};
+
+		let options: cronParser.ParserOptions = {
+			currentDate: formatDate(new Date(), false),
+			tz: 'Asia/Shanghai',
+		};
+
+		try {
+			let interval = cronParser.parseExpression(hoveredWord, options);
+
+			for (let i = 0; i < 10; i++) {
+				data.timeList.push(formatDate(interval.next().toDate()));
+			}
+		} catch (err) {
+			return Result.err(err);
+		}
+		return Result.succ(data);
+	}
+
+	function handleCronTab(hoveredWord: string): vscode.ProviderResult<vscode.Hover> {
+		let markdownArray: Array<vscode.MarkdownString> = [
+			new vscode.MarkdownString('**Conversion CronTab**', true),
+		];
+
+		let convertResult = convertCronTab(hoveredWord);
+
+		// console.log('handleCronTab:', convertResult);
+
+		if (convertResult.isOk()) {
+			markdownArray.push(new vscode.MarkdownString(`Converting ${hoveredWord}`));
+			markdownArray.push(new vscode.MarkdownString('----'));
+
+			const data = convertResult.wrap();
+
+			let text = new vscode.MarkdownString();
+			text.appendMarkdown(`**近${data.timeList.length}次执行时间**\n`);
+			text.appendMarkdown('|序号|执行时间|\n');
+			text.appendMarkdown('|--|--|\n');
+			data.timeList.forEach((it, index) => {
+				text.appendMarkdown(`|${index + 1}|${it}|\n`);
+			});
+			markdownArray.push(text);
+		}
+		return {
+			contents: markdownArray
+		};
+	}
+
+	class DataBase64 {
+		head: string;
+		private code: string = '';
+		private buff: Buffer | undefined = undefined;
+
+		constructor(_head: string = '', _code: string = '') {
+			this.head = _head;
+			this.changeCode(_code);
+		}
+
+		codeDesc(): string {
+			if (this.code.length > 100) {
+				return `${this.code.substring(0, 30)}...${this.code.substring(this.code.length - 30, this.code.length)}`;
+			} else {
+				return this.code;
+			}
+		}
+
+		changeCode(_data: string) {
+			this.code = _data.replace(/((data:\S*;)?base64,)?/g, '');
+			this.buff = Buffer.from(this.code, 'base64');
+		}
+
+		isBase64(): boolean {
+			// Base64 字符串的正则表达式
+			const base64Regex = /^(?:[A-Za-z0-9+/]{4})*(?:[A-Za-z0-9+/]{2}==|[A-Za-z0-9+/]{3}=)?$/g;
+
+			// 先检查字符串长度是否为4的倍数
+			if (!this.code || this.code.length % 4 !== 0) {
+				return false;
+			}
+
+			// 检查字符串是否符合 Base64 字符集规则
+			return base64Regex.test(this.code);
+		}
+
+		isImage(): boolean {
+			return typeof this.head === 'string' && this.head.includes('data:image');
+		}
+
+		toImage(): string | undefined {
+			if (!this.isImage()) {
+				return;
+			}
+			return `${this.head}${this.code}`;
+		}
+
+		isXML(): boolean {
+			const decoded = this.decode();
+			if (decoded) {
+				if (/(>)(<)(\/*)/g.test(decoded)) {
+					return true;
+				}
+			}
+			return false;
+		}
+
+		toXML(): any {
+			if (!this.isXML()) {
+				return;
+			}
+
+			const xmlString = this.decode();
+			if (!xmlString) {
+				return;
+			}
+
+			try {
+				// 将字符串解析为 XML DOM 对象
+				const parser = new DOMParser();
+				const xmlDoc = parser.parseFromString(xmlString, "text/xml");
+
+				// 将 XML DOM 对象序列化为字符串
+				const serializer = new XMLSerializer();
+				const xmlSerialized = serializer.serializeToString(xmlDoc);
+
+				// 使用正则表达式缩进格式化
+				const formatted = xmlSerialized
+					.replace(/(>)(<)(\/*)/g, '$1\r\n$2$3') // 在两个标记之间添加换行符
+					.replace(/(<\w+)(.*?>)/g, function (match, nodeName, nodeProps) {
+						// 缩进层次
+						let indent = match?.match(/<\/?[\w:\-]+>/g)?.length ?? 1;
+						return ' '.repeat(indent * 2) + nodeName + nodeProps;
+					});
+				return formatted;
+			} catch (error) {
+				return error;
+			}
+		}
+
+		decode(): string | undefined {
+			return this.buff?.toString();
+		}
+
+	}
+
+	function convertBase64(hoveredWord: string): Result<DataBase64> {
+		const base64Regex = /((data:image\/svg\+xml;)?base64,)?(?:[A-Za-z0-9+/]{4})*(?:[A-Za-z0-9+/]{2}==|[A-Za-z0-9+/]{3}=)?$/;
+		let data = new DataBase64();
+		hoveredWord.replace(base64Regex, function (_substring, _head): string {
+			// console.log('convertBase64:', _head, ' ---- ', _substring);
+			data.head = _head;
+			data.changeCode(_substring);
+			return '';
+		});
+		// console.log('convertBase64 result:', result);
+		return Result.succ(data);
+	}
+
+	function handleBase64(hoveredWord: string): vscode.ProviderResult<vscode.Hover> {
+		let markdownArray: Array<vscode.MarkdownString> = [
+			new vscode.MarkdownString('**Conversion Base64**', true)
+		];
+
+		const convertResult = convertBase64(hoveredWord);
+
+		// console.log('handleBase64:', convertResult);
+
+		if (!convertResult.isOk()) {
+			return;
+		}
+
+		const data = convertResult.wrap();
+
+		if (!data.isBase64()) {
+			console.error('handleBase64: not Base64!');
+			return;
+		}
+
+		markdownArray.push(new vscode.MarkdownString(`Converting ${data.codeDesc()}`));
+		markdownArray.push(new vscode.MarkdownString('----'));
+
+		let text = new vscode.MarkdownString();
+		text.isTrusted = true;
+		text.supportHtml = true;
+		if (data.isImage()) {
+			markdownArray.push(new vscode.MarkdownString('_IMAGE TYPE_'));
+			text.appendMarkdown(`_${data.head}_`);
+			text.appendMarkdown(`<br><img src="${data.toImage()}"/>`);
+		} else if (data.isXML()) {
+			markdownArray.push(new vscode.MarkdownString('_XML TYPE_'));
+			const xmlResult = data.toXML();
+			if (xmlResult && xmlResult.length) {
+				text.appendCodeblock(xmlResult, 'xml');
+			} else {
+				// text.appendCodeblock( ?? '', 'xml');
+				text.appendText(xmlResult?.toString() ?? '');
+			}
+		} else {
+			markdownArray.push(new vscode.MarkdownString('_UNKNOWN TYPE_'));
+			text.appendCodeblock(data.decode() ?? '');
+		}
+
+		markdownArray.push(text);
+
+		let result: vscode.ProviderResult<vscode.Hover> = {
+			contents: markdownArray
+		};
+		return result;
 	}
 
 	function convertUnicode(hoveredWord: string): string {
@@ -200,11 +512,38 @@ export function activate(context: vscode.ExtensionContext) {
 
 	function handleUnicode(hoveredWord: string): vscode.ProviderResult<vscode.Hover> {
 		let markdownArray: Array<vscode.MarkdownString> = [
-			new vscode.MarkdownString('# Conversion Unicode', true)
+			new vscode.MarkdownString('**Conversion Unicode**', true)
 		];
 
-		const convertedText = convertUnicode(hoveredWord);
-		markdownArray.push(new vscode.MarkdownString(convertedText));
+		let convertedText = convertUnicode(hoveredWord);
+
+		function unicodeDesc(): string {
+			if (hoveredWord.length > 100) {
+				return `${hoveredWord.substring(0, 30)}...${hoveredWord.substring(hoveredWord.length - 30, hoveredWord.length)}`;
+			} else {
+				return hoveredWord;
+			}
+		}
+
+		markdownArray.push(new vscode.MarkdownString(`Converting ${unicodeDesc()}`));
+		markdownArray.push(new vscode.MarkdownString('----'));
+
+		let text = new vscode.MarkdownString();
+		let firstIndex = convertedText.indexOf('{');
+		let lastIndex = convertedText.lastIndexOf('}');
+
+		if (firstIndex !== -1 && lastIndex !== -1) {
+			markdownArray[0] = new vscode.MarkdownString('**Conversion Unicode + JSON**', true);
+			const json = JSON.parse(convertedText.substring(firstIndex, lastIndex + 1));
+
+			text.appendMarkdown(convertedText.substring(0, firstIndex));
+			text.appendCodeblock(JSON.stringify(json, null, 2), 'json');
+			text.appendMarkdown(convertedText.substring(lastIndex + 1, convertedText.length));
+		} else {
+			text.appendText(convertedText);
+		}
+
+		markdownArray.push(text);
 
 		let result: vscode.ProviderResult<vscode.Hover> = {
 			contents: markdownArray
@@ -212,7 +551,12 @@ export function activate(context: vscode.ExtensionContext) {
 		return result;
 	}
 
-	function checkConvertTypeForMultiSource(sources: Array<string>): { convertType: ConvertType, index: number } {
+	interface CheckConvertTypeData {
+		convertType: ConvertType,
+		index: number
+	}
+
+	function checkConvertTypeForMultiSource(sources: Array<string>): CheckConvertTypeData {
 		let convertType: ConvertType = ConvertType.UNKNOWN;
 		let index = 0;
 		for (const it of sources) {
@@ -233,10 +577,44 @@ export function activate(context: vscode.ExtensionContext) {
 		}
 		const timeRegex = /^-?[0-9]+$/g;
 		const unicodeRegex = /\\u([a-fA-F0-9]{4})/g;
+
+		function checkBase64(): boolean {
+			const base64Regex = /(?:[A-Za-z0-9+/]{4})*(?:[A-Za-z0-9+/]{2}==|[A-Za-z0-9+/]{3}=)?$/g;
+
+			if (!base64Regex.test(hoveredWord)) {
+				return false;
+			}
+
+			let splitArr = hoveredWord.split(',');
+			let text = hoveredWord;
+			if (splitArr.length >= 2) {
+				text = splitArr[1];
+			}
+
+			if (text.length % 4 !== 0) {
+				return false;
+			}
+
+			return true;
+		}
+
+		function checkCronTab(): boolean {
+			try {
+				cronParser.parseExpression(hoveredWord);
+				return true;
+			} catch (error) {
+			}
+			return false;
+		}
+
 		if (timeRegex.test(hoveredWord)) {
 			return ConvertType.TIME;
 		} else if (unicodeRegex.test(hoveredWord)) {
 			return ConvertType.UNICODE;
+		} else if (checkCronTab()) {
+			return ConvertType.CRONTAB;
+		} else if (checkBase64()) {
+			return ConvertType.BASE64;
 		}
 		return ConvertType.UNKNOWN;
 	}
@@ -271,9 +649,10 @@ export function activate(context: vscode.ExtensionContext) {
 		var a = parseInt(h, 16);
 		return (a.toString(16) === h.toLowerCase());
 	}
+
 }
 
 // This method is called when your extension is deactivated
 export function deactivate() {
-	console.warn('"preview-conversion" deactivate!');
+	// console.warn('"preview-conversion" deactivate!');
 }